chore: Update @astrojs/mdx configuration
This commit updates the configuration for @astrojs/mdx in the astro.config.mjs file. The syntaxHighlight option is set to "shiki", and the remarkPlugins and rehypePlugins options are updated to include remarkToc and rehypeAccessibleEmojis respectively. Additionally, the shikiConfig option is modified to use the "one-dark-pro" theme and enable line wrapping.