Develop (#46)
* chore(deps): bump @fpkit/react to 0.3.77

updated @fpkit/react package from 0.3.76 to 0.3.77
changed default img component width to 1200 and height to 600

* fix:  remove unused img caption

* refactor: use todo type for collection list and pagination props

- update CollectionList component to use TODO type for collection prop
- update MarkdownPostLayout to specify tag parameter as string
- update [page].astro files to use TODO type for page prop
- import TODO type from types.js in relevant files

* chore(deps): bump @fpkit/react from 0.3.77 to 0.3.78

updated @fpkit/react to version 0.3.78 in package.json, package-lock.json, and yarn.lock files

* docs: components/astro  readme

* docs: update README.md with detailed explanation of Layouts directory

* chore(deps): update astro packages

updated the following astro packages to their latest versions:
- @astrojs/mdx: 2.2.3 -> 2.3.0
- @astrojs/react: 3.1.1 -> 3.3.0
- @astrojs/sitemap: 3.1.1 -> 3.1.2
- astro: 4.5.15 -> 4.6.0
- @astrojs/markdown-remark: 5.0.0 -> 5.1.0
- @astrojs/prism: 3.0.0 -> 3.1.0

this keeps the project dependencies up-to-date with the latest
astro releases and their accompanying improvements, bug fixes,
and potentially new features. staying current helps ensure the
project remains stable, secure, and can take advantage of any
enhancements or optimizations in these updated packages.

* refactor: Update placeholder content in views

- Update content in Reviews, Guides, Patterns, and Placeholder views
- Improve readability and clarity of placeholder text
- Remove unused code and unnecessary metadata

* chore(deps): Update sass dependency to version 1.77.8

* chore: Update @astrojs/mdx to version 2.3.1

* chore: reinstall packages

* chore: Update @astrojs/mdx to version 3.1.3

* chore: Update @astrojs/mdx configuration

This commit updates the configuration for @astrojs/mdx in the astro.config.mjs file. The syntaxHighlight option is set to "shiki", and the remarkPlugins and rehypePlugins options are updated to include remarkToc and rehypeAccessibleEmojis respectively. Additionally, the shikiConfig option is modified to use the "one-dark-pro" theme and enable line wrapping.

* chore: Update @astrojs/mdx configuration and dependencies

This commit updates the configuration for @astrojs/mdx in the astro.config.mjs file. It removes the specific options for syntax highlighting, remark plugins, rehype plugins, and shiki configuration. Additionally, it updates @astrojs/mdx to version 3.1.3 and updates other Astro packages to their latest versions.

* chore: Update @astrojs/mdx configuration and dependencies

* chore: Remove image metadata from blog post files

* chore: Update @astrojs/mdx and @astrojs/netlify dependencies

This commit updates the dependencies for @astrojs/mdx and @astrojs/netlify in the package.json and package-lock.json files. It updates @astrojs/mdx to version 3.1.3 and @astrojs/netlify to version 5.4.0. These updates ensure that the project is using the latest stable versions of these packages.

* chore: Update @fpkit/react dependency to version 0.3.87

* chore: Update @fpkit/react dependency to version 0.3.87

* chore: Update sass dependency to version 1.77.8

* chore: Update sass dependency to version 1.77.8

* chore(deps): bump @fpkit/react from 0.3.89 to 0.3.90

update @fpkit/react dependency in package.json and package-lock.json
- package.json: update version from ^0.3.89 to ^0.3.90
- package-lock.json: update version, resolved url, and integrity hash

---------

Co-authored-by: shawn <shawnsandy04@gmail.com>